# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Muslim Daily API Server",
    "version": "2.0.0",
    "description": "Production-grade RESTful API providing comprehensive Islamic digital services: Holy Quran with audio recitations & translations, accurate prayer timings & trackers, authenticated Hadith collections, authentic Duas & Azkar, multi-system calendar conversions (Gregorian, Hijri, Ethiopian), fasting & journal logs, community challenges, and developer utilities.",
    "contact": {
      "name": "Muslim Daily Engineering Team",
      "email": "abdutahir.dev@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server Instance"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from `/api/auth/login` or `/api/auth/register`."
      }
    },
    "schemas": {
      "DictionaryEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dict_001"
          },
          "word_ar": {
            "type": "string",
            "example": "سَلَام"
          },
          "word_ar_clean": {
            "type": "string",
            "example": "سلام"
          },
          "word_am": {
            "type": "string",
            "example": "ሰላም"
          },
          "word_en": {
            "type": "string",
            "example": "Peace / Greeting"
          },
          "transliteration_ar": {
            "type": "string",
            "example": "Salam"
          },
          "transliteration_am": {
            "type": "string",
            "example": "Selam"
          },
          "part_of_speech": {
            "type": "string",
            "example": "noun"
          },
          "category": {
            "type": "string",
            "example": "Faith & Spirituality"
          },
          "root_ar": {
            "type": "string",
            "example": "س ل م"
          },
          "definition_ar": {
            "type": "string",
            "example": "الأمان والطمأنينة والسكينة وخلو النفس والمجتمع من النزاع."
          },
          "definition_am": {
            "type": "string",
            "example": "መረጋጋት፣ ጸጥታና ፀብ አልባነት፤ እንዲሁም የደህንነት ሰላምታ።"
          },
          "definition_en": {
            "type": "string",
            "example": "A state of harmony, safety, tranquility, and freedom from strife."
          },
          "synonyms": {
            "type": "object",
            "properties": {
              "ar": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "am": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "en": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "antonyms": {
            "type": "object",
            "properties": {
              "ar": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "am": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "en": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ar": {
                  "type": "string"
                },
                "am": {
                  "type": "string"
                },
                "en": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/dictionary": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "List or Search Trilingual Dictionary Entries",
        "description": "Returns paginated dictionary entries with optional search query across Arabic, Amharic, and English, plus filters for category, part of speech, and target language view.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query across Arabic (with/without tashkeel), Amharic (Ge'ez script), English, transliteration, or root"
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "ar",
                "am",
                "en"
              ],
              "default": "all"
            },
            "description": "Focus language formatting"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by thematic category (e.g. Faith & Spirituality, Worship & Law)"
          },
          {
            "name": "pos",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by part of speech (noun, verb, adjective)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "List of dictionary records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DictionaryEntry"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/dictionary/entry/{id}": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Fetch Dictionary Entry by ID",
        "description": "Retrieves a single entry by its unique identifier (e.g., dict_001).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Entry ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Dictionary entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryEntry"
                }
              }
            }
          },
          "404": {
            "description": "Entry not found"
          }
        }
      }
    },
    "/dictionary/daily": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Word of the Day",
        "description": "Returns a deterministic word of the day for spiritual vocabulary enrichment and language learning.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Target date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Daily word",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string"
                    },
                    "word": {
                      "$ref": "#/components/schemas/DictionaryEntry"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}